Fix bug in the bound_form decorator
I recently added the behavior that forms add a class to the html of each
of the form's widgets.  It wasn't catching the exception if a field
wasn't listed in widgets.  As a first fix, I just go to the next field.

Later I'll make it so that it gets the widget for that field (even
though it's not in the widget's dict, and adds it to the widget dict.