fix another bug in form decorator
There is functionality that adds a class name to each widget's html.
For some reason on subsequent calls to the form constructor, in
separate requests, it was appending the class name to whe widgets's
html again and again, accumulating a string of repeating class names.
It seems that the class object is held in memory and persisted, and yet
the decorator gets run over and over, so the default class names
accumulate.

Fixed by checking if the class name is there yet, before appending it.

Also in this commit:
- completed hookup of the Proposal edit view using ProposalVersion form.