Add proposal add form and view
The EditProposalForm has been generalized to also function as an add
form.  It should probably be renamed.  We need selenium testing all over
the proposal editing and adding functionality.

Altered the header's navbar so that the active nav item is determined by
the view context.

Linked 'Issues' nav item to the mainPage (issues listing), and linked '+
create' to the new proposal creation form / view.

Altered mainPage's view slightly so that it doesn't expect an issue
titled "Quebec".