Add AnswerForm, add selenium tests, add ajax error handling
The AnswerForm works and loads up the answer dynamically.  It needs to
be hideable though

I changed how selenium testing works.  I got rid of SeleniumTest as a
test class, and made a module-level DRIVER class, so that the driver can
be shared between all classes.  I made a SeleniumFormTest class that
makes testing forms a bit easier

I added in the ability to generate json_errors to bound_form
functionality, and this is useful for visualizing errors for forms that
are submitted by ajax.

This should be propogated to all ajax endpoints and ajax forms.

the add-letter form (and resend letter form?) currently duplicate
functionality provided by ajax_form (they were a precursor), so they
should use ajax_form for more dryness.