Re-add SeleniumTestCase for live server testing
I had gotten rid of the SeleniumTestCase in favor of creating a
module-wide selenium web-driver that would be shared between
`LiveServerTestCase`s.  But this sharing was causing problems.

I reverted back to the approach of defining a `SeleniumTestCase` which is
an abstract test case which extends `LiveServerTestCase`, and which
provides `driver` and `wait` attributes.  The driver is created and
destroyed for each `SeleniumTestCase`

`SeleniumFormTestCase` once again extends `SeleniumTestCase`.