Make a general comment template
add _w_ajax_form_id.html, which achieves bindings to make form_widgets
in a way that doesn't rely on include_id and is totally driven by the
form object.  This requires the id_prefix of any two forms of the same
type on the same page to be unique.