Add tests for question and answer comments
To do this I added some test answers and comments.  I test both adding
comments to an existing question and answer as well as a newly inserted
answer, because the newly inserted answer (and it's add-comment form)
are dynamically added to the page when the new answer is submitted.

There were some odd issues during testing that went away.  for some
reason pages were taking forever to load.  It seemed like it was because
of a slow connection while loading external scripts, but this doesn't
really make sense because the issues would only happen for the test
server, now when I would navigate to the page manually using the normal
port-8000 server.  I also got strange 'broken pipe' exceptions that,
oddly, would not cause any tests to fail.  It all went away, apparently
when I switched my internet connection.