Create login_required AbstractView and ajax_endpoint decorator.
The AbstractLoginRequiredView derives from AbstractView and adds a check
that the user is logged in arround the view() method, which handles all
requests comming to an AbstractView.

I used this to protect the add_proposal and ask_question views.

I also made a page that the user will be sent to if she attempts to
access a login_required page when she isn't logged in.  It remembers the
page she was trying to access and redirects there after successful
login.  This page is the default place that AbstractLoginRequiredView
sends unauthenticated users, but one can change this by overriding
`get_login_page`.

I made a wrapper for ajax endpoints that should use authentication.
It will do two things.  First, it authenticates the user from which the
request came.  Second, if the request contains a POST variable user, it
will make sure that this is the same authenticated user tied to the
request.  This second part is optional, and the developper must provide
the form class to the decorator for it to be executed.  It's important
because it prevents authenticated users from imporsonating some other
user.

I protected comments and voting in this way.

Also, to prevent surprise from the user, I altered the comment and vote
widgets so that they will be disabled for an unauthenticated user, and a
tooltip appears telling the user they need to login.  This was done
using the jquery tooltip function, which is already in our codebase.

This means that, site-wide, if an html element has a `title` attribute
on it, the text in the title attribute will become a tooltip.  This is
a global behavior, and is switched on inside the head section of the
`__base.html`, currently near line 49.

Now I just need to continue to add authentication required status to the
other views and POSTS around the site that need it, using the
AbstractLoginRequiredView and the `ajax_endpoint_login_required()`
decorator.