Add tooltip and is_enabled attributes to VoteForm
Last commit I was computing wether a vote form should be enabled in the
template.  I've moved this logic into the view.  This is a better place,
because we need to consider both whether the user is logged in, and
whether the user is the author of the content being voted on (users
shouldn't be able to vote on their own content!).

To pass the results of these considerations through to the template,
I bind them to the VoteForm itself.  This involved adding two attributes
to the VoteForm in forms.py.  Corresponding attributes have been added
to the javascript object that handles the state of a vote form, and its
commutication with the server, on the client side.