Add generic AddPost view
AskQuestionView is a class-based view which was basically passing a
QuestionForm to a fairly generic template.  There are several other
cases where we just need to pass a form through to a fairly generic
template.  So, I abstracted out everything that was specificaly
QuestionForm-related, and made an abstract MakePost view.  Now
AskQuestionView just inherets from that and overrides with
Question-specific stuff.  The corresponding template was renamed from
ask_question.html to make_post.html too, and question-related stuff was
stripped out.

Then I subclassed the generic MakePost view to make a
StartDiscussionView.

Some of the patterns used to make `MakePost` and `make_post.html` both
abstract could be applied to other views, which will move some of the
display calculation up to the views, and hopefully decrease the number
of unique templates needed for the site.  That's a TODO for later.